Hotkey troop selection
Pressing Shift+2 puts all selected troops into the "2" hotkey. If you press 2, it will select everything from the 2 hotkey. If you double-tap 2 it will move the camera to one of the soldiers in the 2 hotkey.

ditto for 3,4,5

By default, 1 is the "select all army" hotkey and starts with everything in it